album types
Added methods based on album types.  If an album is a compilation or
soundtrack, it will check to see if there is a separate folder for that
type, then place the files in there by album.  The file names will take
the “artist - song title” form. The singles will also be placed in
their own folder and artist folder.